Fix type mismatch

+++ cores/sprite/HardwareSerial.h
@@ -72,7 +72,7 @@ class HardwareSerial : public Stream
      *
      * @return Amount of bytes transmitted
      */
-    size_t write(uint8_t *buf, unsigned int len);
+    size_t write(uint8_t *buf, uint16_t len);
 
     /**
      * read
